backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.1.0
dnspython==2.4.2
python-dotenv==1.0.0
redis==5.0.1
celery==5.3.4
dkimpy==1.1.8
pyspf==2.0.14
authres==1.2.0
requests==2.31.0
aioredis==2.0.1
asyncpg==0.29.0
httpx==0.25.2
cryptography==41.0.7
PyJWT==2.12.1
bcrypt==4.1.1
starlette==0.27.0
typing-extensions==4.8.0
APScheduler==3.10.4
prometheus-client==0.19.0
structlog==23.2.0
